% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbb-package.R
\docType{package}
\name{bbb-package}
\alias{bbb}
\alias{bbb-package}
\title{bbb: Basic Bounding Box}
\description{
Interactively create a bounding box Drag a
    recrangle, make a bounding box.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/baslat/bbb}
  \item Report bugs at \url{https://github.com/baslat/bbb/issues}
}

}
\keyword{internal}
