% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bollingerBands.R
\name{BollingerBands}
\alias{BollingerBands}
\title{Compute and Disply Bollinger Bands}
\usage{
BollingerBands(instrument)
}
\arguments{
\item{instrument}{A security symbold for which (public) data can be downloaded}
}
\value{
A data.frame with results
}
\description{
This code was written in 2005 in discussion with John Bollinger, and posted for a
few years at the (now long gone) R Graph Gallery.
}
\examples{
\dontrun{
BollingerBands("SPY")
}
}
