% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{box_to_sf}
\alias{box_to_sf}
\title{Convert a drawn box to an SF object}
\usage{
box_to_sf(coords)
}
\arguments{
\item{coords}{from the shiny, the list of coords}
}
\value{
an sf object
}
\description{
Take a rectangle drawn on the leaflet, pull out its coords, and turn it into
a single feature polygon
}
