% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{code_bbox}
\alias{code_bbox}
\title{Get the code to reproduce your bounding box}
\usage{
code_bbox(mbb)
}
\arguments{
\item{mbb}{a manual bounding box from \code{manual_bbox}}
}
\description{
Provides the code to reproduce your bounding box.
}
