% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{manual_bbox}
\alias{manual_bbox}
\title{Manually build a bounding box using the Shiny rectangle}
\usage{
manual_bbox(coords)
}
\arguments{
\item{coords}{from the shiny}
}
\value{
a matrix
}
\description{
Manually construct the bounding box and return a vector.
}
