% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_bbox.R
\name{morph}
\alias{morph}
\title{Morph the format of a bounding box}
\usage{
morph(bbox, output_type)
}
\arguments{
\item{bbox}{a bounding box}

\item{output_type}{(quoted character) the desired output format. One of
\code{"xy_string"}, \code{"word_string"} or \code{"xy_matrix"}.}
}
\value{
a bounding box in the desired format
}
\description{
Change the layout and names of a bounding box. The function will figure out
the current layout of your bounding box and morph it to the desired output
style.
}
\examples{
\dontrun{
bbox <- c("left" = -73.848849, "bottom" = -40.979898, "right" = -28.11848, "top" = 0.703107)
attr(bbox, "class") <- "bbox"
bbox2 <- morph(bbox, "xy_string")
}

}
