% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_bbox.R
\name{to_word_str}
\alias{to_word_str}
\title{Morph any bounding box to a word string bounding box}
\usage{
to_word_str(bbox)
}
\arguments{
\item{bbox}{a bounding box}
}
\value{
a word string bounding box
}
\description{
Take any bounding box format and morph it to a word string type.
}
