% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_bbox.R
\name{word_str_to_xy_matrix}
\alias{word_str_to_xy_matrix}
\title{Morph an xy string bbox to a xy matrix bbox}
\usage{
word_str_to_xy_matrix(bbox)
}
\arguments{
\item{bbox}{A bounding box with names xmin, ymax etc}
}
\value{
a bounding box in matrix form
}
\description{
Changes the format of a xy string bounding box (eg xmax, ymin) to an
xy matrix type.
}
