% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_bbox.R
\name{word_str_to_xy_str}
\alias{word_str_to_xy_str}
\title{Morph a word string bbox to a xy string bbox}
\usage{
word_str_to_xy_str(bbox)
}
\arguments{
\item{bbox}{A bounding box with names left, right etc}
}
\value{
a bounding box uses \code{xmax} etc style as a string
}
\description{
Changes the format of a word string bounding box (ie left, right etc) to an
xy string type (eg xmax, ymin).
}
