% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_bbox.R
\name{xy_str_to_word_str}
\alias{xy_str_to_word_str}
\title{Morph a xy string bbox to a word string bbox}
\usage{
xy_str_to_word_str(bbox)
}
\arguments{
\item{bbox}{A bounding box with names left, right etc}
}
\value{
a bounding box uses \code{xmax} etc style as a string
}
\description{
Changes the format of a xy string bounding box (eg xmax, ymin) to an
word string type (ie left, right etc).
}
