% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbb.R
\name{box}
\alias{box}
\title{Build a bounding box}
\usage{
box()
}
\description{
Launches an interactive \code{shiny} session for you to draw a bounding box.
The code to create the bounding box is then available in your script, or an
object can be created in your session, or both.
}
